Status guage added and pages folder moved
