+++ src/pages/addUser.js
@@ -3,7 +3,7 @@ import Form from "react-validation/build/form";
 import Input from "react-validation/build/input";
 import CheckButton from "react-validation/build/button";
 
-import UserService from "../../service/userService";
+import UserService from "../service/userService";
 
 function required(value) {
   if (!value) {
