+++ src/pages/login.js
@@ -3,7 +3,7 @@ import Form from "react-validation/build/form";
 import Input from "react-validation/build/input";
 import CheckButton from "react-validation/build/button";
 
-import AuthService from "../../service/authService";
+import AuthService from "../service/authService";
 
 const required = (value) => {
   if (!value) {
