+++ src/pages/profile.js
@@ -1,5 +1,5 @@
 import React from "react";
-import AuthService from "../../service/authService";
+import AuthService from "../service/authService";
 
 const Profile = (props) => {
   const currentUser = AuthService.getCurrentUser();
