+++ src/style/gauge.module.css
@@ -1,0 +1,89 @@
+.gauge {
+    width: 80%;
+    height: 80%;
+    display: flex;
+    border-radius: 1%;
+}
+
+.gaugeItem {
+    width: 10%;
+    border-radius: 5px;
+    background-color: blue;
+    box-shadow: inset 5px -5px 10px 2px rgba(0, 0, 116, 0.705),
+        2px 2px 10px 3px rgba(0, 0, 0, 0.596);
+    display: flex;
+    align-items: center;
+    justify-content: center;
+}
+
+.gaugeItem:nth-child(n+4):nth-child(-n+6){
+    width: 14%;
+}
+
+.gaugeItem:nth-child(7){
+    width: 16%;
+}
+
+.gaugeItem:last-child {
+    width: 16%;
+    box-shadow: 0px 0px 20px 5px rgba(132, 132, 248, 0.986),
+        inset 5px -5px 10px 2px rgba(0, 0, 116, 0.705);
+    transform: scale(1.1);
+    text-shadow: 0px 0px 15px white,
+        0px 0px 15px rgba(132, 132, 248, 0.986);
+
+
+
+
+}
+
+.gaugeNumber {
+    font-size: 3rem;
+    color: white;
+
+}
+
+.gaugeCancel {
+    width: 80%;
+    height: 80%;
+    border-radius: 5px;
+    background-color: red;
+    box-shadow: 0px 0px 20px 5px rgba(241, 106, 106, 0.822),
+        inset 5px -5px 10px 2px rgba(126, 6, 6, 0.76);
+    text-align: center;
+}
+
+.cancel {
+    width: 100%;
+    font-size: 3rem;
+    color: white;
+    text-shadow: 0px 0px 15px white,
+        0px 0px 15px rgba(132, 132, 248, 0.986);
+
+
+}
+
+.meter {
+    width: 70vw;
+    height: 7vh;
+}
+
+
+/*====================================================
+  ==                                                ==
+  ==                                                ==
+  ==             Mobile Styling                     ==
+  ==                                                ==
+  ==                                                ==
+  ====================================================*/
+
+@media only screen and (max-width: 550px) {
+
+    .gaugeItem {
+        flex-direction: column;
+    }
+
+    .gaugeNumber {
+        font-size: 2rem;
+    }
+}
