+++ src/style/orders.module.css
@@ -3,6 +3,7 @@
   display: flex;
   flex-direction: column;
   align-items: center;
+  
 }
 
 .inputContainer {
@@ -34,14 +35,14 @@
 
 .flagItem {
   width: 50vw;
-  min-height: 20vh;
+  min-height: 30vh;
   position: relative;
   border-radius: 10px;
   margin-bottom: 4vh;
   box-shadow: rgba(0, 0, 0, 0.3) 0px 15px 20px, rgba(0, 0, 0, 0.22) 0px 11px 12px;
   display: grid;
   grid-template-columns: 60% 40%;
-  grid-template-rows: 20% 80%;
+  grid-template-rows: 20% 50% 30%;
 }
 
 
@@ -80,7 +81,7 @@
 }
 
 .description {
-  font-size: 2.7rem;
+  font-size: 2.4rem;
   font-weight: 700;
   text-align: center;
 }
@@ -137,6 +138,13 @@
 
 }
 
+.gaugeContainer{
+  grid-row: 3/4;
+  grid-column: 1/3;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
 
 /*====================================================
   ==                                                ==
@@ -146,7 +154,7 @@
   ==                                                ==
   ====================================================*/
 
-@media only screen and (min-width: 550px) and (max-width: 950px) {
+@media only screen and (min-width: 550px) and (max-width: 987px) {
 
 
   .flagItem {
@@ -205,6 +213,13 @@
   .statusGroup {
     order: 3;
   }
+
+  .gaugeContainer{
+    order: 5;
+    width: 100%;
+  }
+
+  
 
   .statusItemContainer {
     display: none;
